Add support for `CARGO_MOMMYS_ACTUAL` to control child program
It may potentially be desirable to use mommy with not-cargo programs.
While today we can use `CARGO` to control this, this does set the
`CARGO` env var, which may make some tools unhappy (for example, if
another cargo gets called beneath this other program). To avoid this,
this adds a neutral environment variable, `CARGO_MOMMYS_ACTUAL`.